% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{create_key_from_rubric}
\alias{create_key_from_rubric}
\title{Create \code{psych} scoring key from tibble of scoring rubrics.}
\usage{
create_key_from_rubric(rubricsDF, scale_name = NULL)
}
\arguments{
\item{rubricsDF}{A long-format tibble of scoring rubrics as returned by \code{\link{get_rubrics}}.}

\item{scale_name}{Name of the scale you wish to create a scoring key for, as a character string.
By default, all scales represented in \code{rubricsDF} are selected (i.e. \code{scale_name} is set to \code{NULL}).}
}
\value{
A named list containing scoring keys to be used by the \code{psych} package.
}
\description{
Create \code{psych} scoring key from tibble of scoring rubrics.
}
