% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devaluation-package.R
\docType{package}
\name{devaluation-package}
\alias{devaluation}
\alias{devaluation-package}
\title{devaluation: Access data from the Devaluation Study from UO SAN Lab}
\description{
This package contains a set of functions that facilitate API 
    requests to Qualtrics and Redcap in order to fetch raw data from the 
    Devaluation study, conducted by the Social and Affective Neuroscience Lab 
    at the University of Oregon.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UOSAN/devaluation}
  \item \url{https://uosan.github.io/devaluation/}
  \item Report bugs at \url{https://github.com/UOSAN/devaluation/issues}
}

}
\author{
\strong{Maintainer}: Brendan Cullen \email{bcullen@uoregon.edu}

}
\keyword{internal}
