% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{get_rubrics}
\alias{get_rubrics}
\title{Get scoring rubrics}
\usage{
get_rubrics(path_to_rubrics)
}
\arguments{
\item{path_to_rubrics}{Character string specifying full path to directory where
scoring rubrics are saved. Rubrics should be saved as .csv files and named according
to the following convention: \code{<measure>_scoring_rubric.csv}, e.g. \code{PLAN_scoring_rubric.csv}
for the Planfulness Scale}
}
\value{
Returns a long-format tibble of scoring rubrics with the following column names:
"file"           "data_file_name" "scale_name"     "column_name"
"reverse"        "transform"      "scored_scale"   "include"
"min"            "max".
}
\description{
Wrangle scoring rubrics into a long-format tibble.
}
