% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get survey data}
\usage{
get_survey_data(surveys_df, pid_col = "Login")
}
\arguments{
\item{surveys_df}{A tibble as returned by the \code{\link{get_survey_info}}.}

\item{pid_col}{A character with the column name specifying user ID.}
}
\value{
Returns a long format tibble of survey data with names "SID variable name", "item", "value", "survey_name"
}
\description{
Get survey data
}
