% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_info.R
\name{get_survey_info}
\alias{get_survey_info}
\title{Get survey information.}
\usage{
get_survey_info()
}
\value{
A tibble including the following variables:
\itemize{
\item id
\item name
\item ownerId
\item lastModified
\item creationDate
\item isActive
}

For more information, see \code{\link[qualtRics]{all_surveys}}
}
\description{
A wrapper function for \code{\link[qualtRics]{all_surveys}}. This function returns metadata
for surveys administered at Devaluation sessions (0-5).
}
