% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_responses}
\alias{get_survey_responses}
\title{Return a tibble of the survey responses for a given survey}
\usage{
get_survey_responses(
  survey_id,
  label = FALSE,
  convert = FALSE,
  force_request = TRUE
)
}
\arguments{
\item{survey_id}{Unique survey ID as returned from \code{\link{get_survey_info}}.}

\item{label}{default is \code{FALSE}.}

\item{convert}{default is \code{FALSE}.}

\item{force_request}{default is \code{TRUE}.}
}
\description{
See \code{\link[qualtRics]{fetch_survey}} for complete information on available options.
}
