% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_qualtrics_credentials.R
\name{register_qualtrics_credentials}
\alias{register_qualtrics_credentials}
\title{Register Qualtrics API credentials in your current session.}
\usage{
register_qualtrics_credentials(path_to_creds)
}
\arguments{
\item{path_to_creds}{Full path to credentials file containing API token and base URL.
The credentials file must have the following columns: data_source, base_url, api_token.}
}
\description{
Reads the API token and base URL from a credentials file (in .csv format) and sets up the environment
variables necessary for accessing data. Uses
\code{\link[qualtRics]{qualtrics_api_credentials}}, but does not save
credentials to the .Renviron file.
}
