% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devaluation-package.R
\docType{package}
\name{devaluation-package}
\alias{devaluation}
\alias{devaluation-package}
\title{devaluation: Access data from the Devaluation Study from UO SAN Lab}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brendan Cullen \email{bcullen@uoregon.edu}

}
\keyword{internal}
