% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_tools.R
\name{get_rubrics}
\alias{get_rubrics}
\title{Get rubrics}
\usage{
get_rubrics(path_to_rubrics)
}
\arguments{
\item{path_to_rubrics}{Character string specifying full path to directory where
scoring rubrics are saved.}
}
\value{
Returns a long data frame of rubrics with names:
"file"           "data_file_name" "scale_name"     "column_name"
"reverse"        "transform"      "scored_scale"   "include"
"min"            "max".
}
\description{
Get rubrics
}
