% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_info.R
\name{get_survey_info}
\alias{get_survey_info}
\title{Get Qualtrics survey info}
\usage{
get_survey_info(append_raw_data = FALSE)
}
\arguments{
\item{append_raw_data}{A logical value indicating whether raw survey data should
be included as a column in the output}
}
\value{
A tibble with the following columns:
\itemize{
\item id (unique survey identifier)
\item name (survey name)
\item lastModified (date of most recent response)
\item n_obs (number of observations in raw survey data)
}

If \code{append_raw_data} is set to \code{TRUE}, output includes an additional
list column, survey_data.
}
\description{
Get information about all DEV Qualtrics surveys, including number of
observations in raw data and date of most recent response
}
